img-56 Repackage Security Classes
This change creates a org.codice.nitf.imaging.core.security package and moves the security-related classes into it.
