+++ core/src/main/java/org/codice/imaging/nitf/core/NitfConstants.java
@@ -304,261 +304,6 @@ public final class NitfConstants {
      */
     protected static final long STREAMING_FILE_MODE = 999999999999L;
 
-    // file security
-    /**
-     * Length of the "File Copy Number" field in the NITF file header.
-     * <p>
-     * See MIL-STD-2500C Table A-1.
-     */
-    protected static final int FSCOP_LENGTH = 5;
-
-    /**
-     * Length of the "File Number of Copies" field in the NITF file header.
-     * <p>
-     * See MIL-STD-2500C Table A-1.
-     */
-    protected static final int FSCPYS_LENGTH = 5;
-
-    // security metadata
-    /**
-     * Length of the "Security Classification" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSCLAS_LENGTH = 1;
-
-    /**
-     * Length of the "Security Classification System" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSCLSY_LENGTH = 2;
-
-    /**
-     * Length of the "Codewords" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-
-    protected static final int XSCODE_LENGTH = 11;
-
-    /**
-     * Length of the "Control and Handling" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSCTLH_LENGTH = 2;
-
-    /**
-     * Length of the "Releasing Instructions" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSREL_LENGTH = 20;
-
-    /**
-     * Length of the "Declassification Type" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSDCTP_LENGTH = 2;
-
-    /**
-     * Length of the "Declassification Date" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSDCDT_LENGTH = 8;
-
-    /**
-     * Length of the "Declassification Exemption" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSDCXM_LENGTH = 4;
-
-    /**
-     * Length of the "Downgrade" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSDG_LENGTH = 1;
-
-    /**
-     * Length of the "Downgrade Date" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSDGDT_LENGTH = 8;
-
-    /**
-     * Length of the "Classification Text" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSCLTX_LENGTH = 43;
-
-    /**
-     * Length of the "Classification Authority Type" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSCATP_LENGTH = 1;
-
-    /**
-     * Length of the "Classification Authority" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSCAUT_LENGTH = 40;
-
-    /**
-     * Length of the "Classification Reason" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSCRSN_LENGTH = 1;
-
-    /**
-     * Length of the "Security Source Date" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSSRDT_LENGTH = 8;
-
-    /**
-     * Length of the "Security Control Number" field used in the NITF file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500C Annex A for the various uses.
-     */
-    protected static final int XSCTLN_LENGTH = 15;
-
-    // NITF 2.0 field lengths
-    /**
-     * Length of the "Codewords" field used in the NITF 2.0 file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500A Tables for the various uses.
-     */
-    protected static final int XSCODE20_LENGTH = 40;
-
-    /**
-     * Length of the "Control and Handling" field used in the NITF 2.0 file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500A Tables for the various uses.
-     */
-    protected static final int XSCTLH20_LENGTH = 40;
-
-    /**
-     * Length of the "Releasing Instructions" field used in the NITF 2.0 file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500A Tables for the various uses.
-     */
-    protected static final int XSREL20_LENGTH = 40;
-
-    /**
-     * Length of the "Classification Authority" field used in the NITF 2.0 file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500A Tables for the various uses.
-     */
-    protected static final int XSCAUT20_LENGTH = 20;
-
-    /**
-     * Length of the "Control Number" field used in the NITF 2.0 file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500A Tables for the various uses.
-     */
-    protected static final int XSCTLN20_LENGTH = 20;
-
-    /**
-     * Length of the "Security Downgrade" field used in the NITF 2.0 file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500A Tables for the various uses.
-     */
-    protected static final int XSDWNG20_LENGTH = 6;
-
-    /**
-     * Length of the "Downgrade Event" field used in the NITF 2.0 file header and subheaders.
-     * <p>
-     * This field is called different things in different places. For example, in the file header
-     * it starts with F, while in the image subheader it starts with I. To make it more general,
-     * we use 'X' to represent these various prefixes.
-     * See MIL-STD-2500A Tables for the various uses.
-     */
-    protected static final int XSDEVT20_LENGTH = 40;
-
-    /**
-     * Marker field used in the "Security Downgrade" field to indicate that the "Downgrade Event" field is present.
-     * <p>
-     * NITF 2.0 only. See MIL-STD-2500A Tables for usage.
-     */
-    public static final String DOWNGRADE_EVENT_MAGIC = "999998";
-
     // text segment
     /**
      * Marker string for NITF Text subheader.
@@ -633,6 +378,13 @@ public final class NitfConstants {
 
     static final String AND_CONDITION = " AND ";
     static final String UNSUPPORTED_IFTYPE_FORMAT_MESSAGE = "Unsupported format for iftype:";
+
+    /**
+     * Marker field used in the "Security Downgrade" field to indicate that the "Downgrade Event" field is present.
+     * <p>
+     * NITF 2.0 only. See MIL-STD-2500A Tables for usage.
+     */
+    public static final String DOWNGRADE_EVENT_MAGIC = "999998";
 
     private NitfConstants() {
     }
