+++ core/src/test/java/org/codice/imaging/nitf/core/Nitf20SymbolTest.java
@@ -14,13 +14,13 @@
  **/
 package org.codice.imaging.nitf.core;
 
-import java.io.BufferedInputStream;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 
-import java.io.InputStream;
+import java.io.BufferedInputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.TimeZone;
@@ -28,6 +28,9 @@ import java.util.TimeZone;
 import org.codice.imaging.nitf.core.common.FileType;
 import org.codice.imaging.nitf.core.common.NitfInputStreamReader;
 import org.codice.imaging.nitf.core.common.NitfReader;
+import org.codice.imaging.nitf.core.security.FileSecurityMetadata;
+import org.codice.imaging.nitf.core.security.SecurityClassification;
+import org.codice.imaging.nitf.core.security.SecurityMetadata;
 import org.junit.Before;
 import org.junit.Test;
