+++ core/src/test/java/org/codice/imaging/nitf/core/graphic/Nitf21GraphicParsingTest.java
@@ -26,8 +26,8 @@ import org.codice.imaging.nitf.core.HeaderOnlyNitfParseStrategy;
 import org.codice.imaging.nitf.core.NitfFileParser;
 import org.codice.imaging.nitf.core.common.NitfInputStreamReader;
 import org.codice.imaging.nitf.core.common.NitfReader;
-import org.codice.imaging.nitf.core.SecurityClassification;
-import org.codice.imaging.nitf.core.SecurityMetadata;
+import org.codice.imaging.nitf.core.security.SecurityClassification;
+import org.codice.imaging.nitf.core.security.SecurityMetadata;
 import org.junit.Test;
 
 public class Nitf21GraphicParsingTest {
