+++ core/src/test/java/org/codice/imaging/nitf/core/image/NitfImageSegmentHeaderParserTest.java
@@ -25,8 +25,8 @@ import java.util.LinkedList;
 import org.codice.imaging.nitf.core.common.FileType;
 import org.codice.imaging.nitf.core.common.NitfParseStrategy;
 import org.codice.imaging.nitf.core.common.NitfReader;
-import org.codice.imaging.nitf.core.SecurityClassification;
-import org.codice.imaging.nitf.core.SecurityMetadata;
+import org.codice.imaging.nitf.core.security.SecurityClassification;
+import org.codice.imaging.nitf.core.security.SecurityMetadata;
 import org.codice.imaging.nitf.core.PixelJustification;
 import org.codice.imaging.nitf.core.PixelValueType;
 import org.junit.Before;
